refactor: badge helper
